requirements.txt
numpy
pandas
geopandas
torch
segmentation_models_pytorch
torchvision
torchinfo
transformers
tensorboard
torchmetrics
matplotlib
grequests
planetary-computer
pystac
pystac-client
stackstac
ipyleaflet
rich
haversine
dask[complete]
shapely
rasterio
rioxarray
xarray-spatial
geocube
ipympl
opencv-python-headless
empatches
richdem
loguru
albumentations
jupyter
ipykernel
pygeos
fiona